# Faster deserialization by loading BitmapContainer in a single .readFully