--- a/jmh/src/main/java/org/roaringbitmap/realdata/RealDataSerializationBenchmark.java
+++ b/jmh/src/main/java/org/roaringbitmap/realdata/RealDataSerializationBenchmark.java
@@ -40,6 +40,11 @@ import org.roaringbitmap.RoaringBitmap;
 import org.roaringbitmap.ZipRealDataRetriever;
 import org.roaringbitmap.buffer.ImmutableRoaringBitmap;
 
+/**
+ * 
+ * @author Richard Startin
+ *
+ */
 public class RealDataSerializationBenchmark {
 
 	 @Test