… examples/SerializeToByteArrayExample.java
… jmh/pom.xml
… pom.xml
… roaringbitmap/src/main/java/SerializeToByteArrayExample.java
… roaringbitmap/src/main/java/org/roaringbitmap/RoaringBitmap.java
… roaringbitmap/src/main/java/org/roaringbitmap/buffer/ImmutableRoaringBitmap.java
… roaringbitmap/src/main/java/org/roaringbitmap/insights/BitmapAnalyser.java
… roaringbitmap/src/main/java/org/roaringbitmap/insights/BitmapStatistics.java
… roaringbitmap/src/main/java/org/roaringbitmap/insights/NaiveWriterRecommender.java
… roaringbitmap/src/test/java/org/roaringbitmap/TestRoaringBitmap.java
… roaringbitmap/src/test/java/org/roaringbitmap/TestSerialization.java
… roaringbitmap/src/test/java/org/roaringbitmap/TestSerializationViaByteBuffer.java
… roaringbitmap/src/test/java/org/roaringbitmap/longlong/Issue319.java