The IO effect of reading bytes one by one seems quite big. With a single .readFully:

recommended: 
247 ms 107 ms 83 ms 177 ms 81 ms 81 ms 107 ms 83 ms 83 ms 116 ms 154 ms 97 ms 82 ms 85 ms 83 ms 84 ms 97 ms 84 ms 84 ms 83 ms 83 ms 117 ms 161 ms 132 ms 82 ms 84 ms 84 ms 84 ms 84 ms 82 ms 
average: 103
via ByteArrayInputStream: 
257 ms 86 ms 86 ms 86 ms 86 ms 84 ms 82 ms 82 ms 180 ms 88 ms 83 ms 85 ms 84 ms 84 ms 85 ms 86 ms 102 ms 90 ms 84 ms 82 ms 83 ms 84 ms 83 ms 85 ms 94 ms 82 ms 81 ms 82 ms 84 ms 82 ms 
average: 94
via Immutable: 
130 ms 69 ms 77 ms 67 ms 67 ms 70 ms 67 ms 64 ms 62 ms 64 ms 77 ms 65 ms 64 ms 63 ms 64 ms 64 ms 64 ms 62 ms 66 ms 64 ms 66 ms 67 ms 64 ms 65 ms 64 ms 62 ms 78 ms 66 ms 65 ms 64 ms 
average: 68

Reference: https://github.com/RoaringBitmap/RoaringBitmap/issues/319