@blacelle 2019-03-15
leaving little-endian management to byteBuffer:

recommended: 
256 ms 81 ms 77 ms 142 ms 75 ms 75 ms 95 ms 74 ms 75 ms 99 ms 138 ms 77 ms 73 ms 74 ms 76 ms 131 ms 75 ms 74 ms 74 ms 74 ms 74 ms 133 ms 170 ms 74 ms 76 ms 80 ms 77 ms 77 ms 78 ms 108 ms 
average: 95
via ByteArrayInputStream: 
102 ms 100 ms 99 ms 75 ms 80 ms 76 ms 79 ms 82 ms 77 ms 76 ms 78 ms 78 ms 78 ms 76 ms 85 ms 76 ms 76 ms 79 ms 79 ms 77 ms 76 ms 102 ms 76 ms 77 ms 75 ms 75 ms 76 ms 75 ms 76 ms 79 ms 
average: 80
via Immutable: 
91 ms 69 ms 69 ms 68 ms 66 ms 64 ms 77 ms 67 ms 65 ms 62 ms 62 ms 63 ms 65 ms 83 ms 62 ms 63 ms 62 ms 66 ms 66 ms 73 ms 67 ms 67 ms 63 ms 63 ms 66 ms 68 ms 63 ms 65 ms 70 ms 64 ms 
average: 67