@blacelle 2019-03-15
Same pattern to other containers does not improve performance on Issue319 gist . But I guess this is because Issue319 gist generate mainly BitmapContainers