@lemire 2019-03-15
That's nice work. I guess that there is no catching up to "via Immutable" because we don't get the array directly, so there is extra monkeying around.