@blacelle 2019-03-15
Given code is not ready: for RunContainers and ArrayContainers, I guessed the buffer for BitmapContainer was big enough. I do not know enough constrains on these containers to know if the constraint is relyable, or if a growing buffer should be considered.

There is also org.roaringbitmap.KryoTest.roaringTest() which fails, but I can not find why right now.