@blacelle 2019-03-15
With RoaringBitmap.deserialize(ByteBuffer)

recommended: 
256 ms 82 ms 75 ms 143 ms 74 ms 73 ms 97 ms 74 ms 76 ms 99 ms 136 ms 80 ms 74 ms 76 ms 76 ms 137 ms 74 ms 74 ms 75 ms 75 ms 74 ms 129 ms 192 ms 74 ms 75 ms 77 ms 81 ms 83 ms 83 ms 108 ms 
average: 96
via ByteArrayInputStream: 
102 ms 101 ms 91 ms 76 ms 77 ms 77 ms 78 ms 79 ms 76 ms 75 ms 76 ms 80 ms 78 ms 75 ms 80 ms 76 ms 75 ms 75 ms 75 ms 76 ms 75 ms 84 ms 75 ms 74 ms 76 ms 76 ms 77 ms 75 ms 89 ms 75 ms 
average: 79
via Immutable: 
94 ms 67 ms 66 ms 65 ms 63 ms 67 ms 66 ms 62 ms 62 ms 65 ms 63 ms 64 ms 65 ms 71 ms 68 ms 67 ms 63 ms 66 ms 61 ms 63 ms 88 ms 65 ms 66 ms 65 ms 63 ms 66 ms 63 ms 64 ms 67 ms 71 ms 
average: 66
via Deserialize ByteBuffer: 
83 ms 71 ms 62 ms 59 ms 62 ms 93 ms 61 ms 60 ms 59 ms 60 ms 59 ms 60 ms 60 ms 73 ms 60 ms 60 ms 60 ms 60 ms 61 ms 60 ms 59 ms 65 ms 61 ms 61 ms 67 ms 62 ms 60 ms 59 ms 64 ms 62 ms 
average: 63