@richardstartin 2019-03-16
I hadn't realised that this PR already had a `ByteBuffer` implementation virtually the same as in #321, and this one was first so I will close that PR. However, I have reservations about the changes made to the method which uses a `DataInput` here. They are made on the basis of observations about one implementation of `DataInput`, which don't generalise to other implementations see [the comment with benchmarks on the issue](https://github.com/RoaringBitmap/RoaringBitmap/issues/319#issuecomment-473516728) . They also require that potentially very large (tens of megabytes) `byte[]` arrays are allocated. I would like to see this PR merged, but would, as a user of the method being modified, first like to see:

* The changes to `serialize(DataInput)` reverted
* The new method `serialize(ByteBuffer)` kept and promoted as the most efficient option, and tested with a range of byte orders and buffer varieties
* A proper benchmark, using JMH, in the benchmarks project, **not** in the tests.