@blacelle 2019-03-16
Perf Comparison with #321 

recommended: 
282 ms 89 ms 81 ms 193 ms 83 ms 82 ms 97 ms 75 ms 77 ms 109 ms 153 ms 89 ms 78 ms 81 ms 77 ms 78 ms 83 ms 75 ms 76 ms 75 ms 79 ms 123 ms 143 ms 93 ms 76 ms 79 ms 78 ms 78 ms 77 ms 78 ms 
average: 97
via ByteArrayInputStream: 
135 ms 99 ms 90 ms 75 ms 77 ms 77 ms 78 ms 84 ms 114 ms 76 ms 79 ms 79 ms 78 ms 78 ms 76 ms 78 ms 81 ms 78 ms 76 ms 76 ms 77 ms 77 ms 77 ms 78 ms 104 ms 78 ms 78 ms 76 ms 78 ms 77 ms 
average: 82
via Immutable: 
92 ms 68 ms 73 ms 64 ms 65 ms 65 ms 64 ms 65 ms 63 ms 65 ms 97 ms 65 ms 61 ms 63 ms 62 ms 63 ms 65 ms 65 ms 102 ms 66 ms 63 ms 64 ms 64 ms 65 ms 66 ms 66 ms 63 ms 66 ms 65 ms 66 ms 
average: 68
via Deserialize ByteBuffer: 
69 ms 71 ms 61 ms 58 ms 60 ms 67 ms 63 ms 62 ms 57 ms 60 ms 59 ms 59 ms 60 ms 58 ms 61 ms 59 ms 61 ms 59 ms 61 ms 62 ms 61 ms 61 ms 73 ms 62 ms 61 ms 63 ms 62 ms 63 ms 61 ms 59 ms 
average: 61
via Deserialize ByteBuffer (321): 
90 ms 73 ms 89 ms 65 ms 68 ms 66 ms 65 ms 68 ms 69 ms 77 ms 65 ms 66 ms 66 ms 65 ms 65 ms 66 ms 68 ms 67 ms 69 ms 66 ms 67 ms 72 ms 66 ms 66 ms 68 ms 65 ms 76 ms 68 ms 69 ms 66 ms 
average: 69