@richardstartin 2019-03-16
The implementations are virtually identical so I would be surprised if there were any difference that couldn't be explained away as noise, and I'm not advocating for the use of the version in #321 because it's virtually the same thing. 

However, the way you are measuring performance is flawed - could you write a benchmark using JMH to make sure the benchmarks run in forked JVMs, warm up properly, measure over a range of inputs, and so on?