@blacelle 2019-03-16
> They also require that potentially very large (tens of megabytes) byte[] arrays are allocated

I suggest adjusting the implementation to rely on a single byte[] based on BitmapContainer size. For the other containers, if the buffer is too small, it could be used to load by chunks (instead of 1 by 1 in original implementation, or all in a single byte[] in current proposition).

@richardstartin Yes, I feel the simpler would be to port the JMH benchmark from #320 . For now, the original gist was enough to pin the performance issue.