@richardstartin 2019-03-16
Let me outline again why I would prefer that this PR does not modify the old method and simply introduces the new method:

- The benchmark evidence is based on `DataInputStream`, not `RandomAccessFile` etc. on a server class machine (i.e. not on a probably encrypted HDD). I think the benchmark evidence actually just highlights how bad `DataInputStream`/`ByteArrayInputStream` are.
- The modified method used to create very little temporary memory but now it creates a lot of it, even if you allocate just 8kB as you suggest. This may be an issue for users when picking up the new version. @ppiotrow has previously claimed to have 5 million bitmaps in his application - this would imply 40GB temporary memory to deserialise his bitmaps, assuming the change to use a 8kB buffer is made, and may require a buffering solution... 
- The new method which takes a `ByteBuffer` as input is just a much, much better solution, and supports something similar to the previous `RandomAccessFile` use case simply via `MappedByteBuffer`. In fact, it's exactly the buffering solution you would need to make point 2 work.

@blacelle I would prefer that you do not disregard my concerns as a user because an increase in memory consumption may cause me problems, but do note that I'm looking forward to having access to the new method.

@lemire please take my concerns as a user into account, and I suggest it may be worthwhile to collect feedback from other users too, given the space-time trade off being proposed here, before merging the modified method.