@blacelle 2019-03-17
@richardstartin Nobody disregards your feedback. This PR has been open to showcase one way to address the original penalty. For now, I feel nothing is decided regarding `org.roaringbitmap.RoaringArray.deserialize(DataInput)`