@blacelle 2019-03-17
About @ppiotrow use-case, I wonder if it would rely on .deserialize(ByteBuffer), or if the GC impact of the buffer would be counter-balancer by the performance issue revealed in #319