@richardstartin 2019-03-17
The point I’m trying to make is that isn’t your call to make. This method is being used, and you want to make a space time trade off on the behalf of users. If this is merged, I will need to undertake unplanned work if I update for a bug fix. Why do you want to make this change? Because it affects your application? Or because a performance issue was raised and you want to fix it? If it’s the latter, I reiterate that introducing the new method and changing the documentation is the right course of action.